requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
narwhals==1.35.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
plotly==6.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.21.0
python-dotenv
